Properly project layers
One layer is the base layer, and every other layer is projected in the
CRS of the base layer. This makes points on different layers line up.
Currently only base layers with the unit of meters are supported.

It is possibly to make raster layers non base layers, but it would
require reprojecting the tiles. For now that isn't available.

See <https://openlayers.org/en/latest/doc/tutorials/raster-reprojection.html>